# swagger.yml
openapi: 3.0.3
info:
  title: EduQuest
  description: |
    Приложение для улучшения качества образования с элементами геймификации.
    **Основные возможности:**
    - Анонимная обратная связь после пар
    - Вопросы во время занятий
    - Викторины с начислением баллов
    - Карточки для повторения
    - Тренажёр решения задач
    - Возможность обмена баллов на награды
    **Роли пользователей:**
    - student — студент
    - teacher — преподаватель
    - admin — администратор
  version: 1.0.0
  contact:
    name: EduQuest Team
    email: support@eduquest.ru
servers:
  - url: http://localhost/eduquest
    description: Локальный сервер разработки
  - url: https://api.eduquest.ru
    description: Продакшен сервер
...
paths:
...
  /api/groups:
    post:
      tags: [Groups]
      summary: Создать новую учебную группу
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
      responses:
        '201':
          description: Группа создана
        '400':
          description: Некорректные данные
        '403':
          description: Только администраторы могут создавать группы
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера
    get:
      tags: [Groups]
      summary: Получить список учебных групп
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Список групп
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера

  # Карточки
  /api/cards:
    post:
      tags: [Cards]
      summary: Создать новую карточку
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Card'
      responses:
        '201':
          description: Карточка создана
        '400':
          description: Некорректные данные
        '403':
          description: Только преподаватели могут создавать карточки
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера
    get:
      tags: [Cards]
      summary: Получить карточки для повторения
      security:
        - BearerAuth: []
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: topic
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список карточек
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера

  # Задачи
...
  /api/groups/{group_id}/questions:
    post:
      tags: [Questions]
      summary: Создать вопрос для группы
      security:
        - BearerAuth: []
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Question'
      responses:
        '201':
          description: Вопрос создан
        '400':
          description: Некорректные данные
        '403':
          description: Только преподаватели могут создавать вопросы
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера
    get:
      tags: [Questions]
      summary: Получить вопросы для группы
      security:
        - BearerAuth: []
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список вопросов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
        '401':
          description: Требуется аутентификация
        '403':
          description: Доступ запрещён
        '500':
          description: Ошибка сервера
  /api/questions/{id}:
    delete:
      tags: [Questions]
      summary: Удалить вопрос
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '204':
          description: Вопрос удалён
        '403':
          description: Только автор может удалить вопрос
        '404':
          description: Вопрос не найден
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера
  /api/questions/{question_id}/answers:
    post:
      tags: [Questions]
      summary: Ответить на вопрос
      security:
        - BearerAuth: []
      parameters:
        - name: question_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Answer'
      responses:
        '201':
          description: Ответ добавлен
        '400':
          description: Некорректные данные
        '403':
          description: Только студенты могут отвечать
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера
    get:
      tags: [Questions]
      summary: Получить ответы на вопрос
      security:
        - BearerAuth: []
      parameters:
        - name: question_id
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список ответов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Answer'
        '403':
          description: Доступ запрещён
        '401':
          description: Требуется аутентификация
        '500':
          description: Ошибка сервера

  # Обратная связь
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        first_name:
          type: string
          example: "Иван"
        last_name:
          type: string
          example: "Иванов"
        email:
          type: string
          format: email
          example: "ivan@university.edu"
        role:
          type: string
          enum: [student, teacher, admin]
          example: "student"
        points:
          type: integer
          example: 150
        group:
          $ref: '#/components/schemas/Group'
        department:
          type: string
          example: "Математический факультет"
      required: [id, first_name, last_name, email, role, group]
    Group:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "МАТ-101"
        faculty:
          type: string
          example: "Математический факультет"
      required: [id, name]
    Question:
      type: object
      properties:
        id:
          type: integer
          example: 1
        text:
          type: string
          example: "Можете повторить доказательство теоремы?"
        teacher:
          $ref: '#/components/schemas/User'
        group:
          $ref: '#/components/schemas/Group'
        created_at:
          type: string
          format: date-time
        is_active:
          type: boolean
          example: true
      required: [text, teacher, group]
    Answer:
      type: object
      properties:
        id:
          type: integer
          example: 1
        text:
          type: string
          example: "Конечно, вот доказательство..."
        student:
          $ref: '#/components/schemas/User'
        question:
          $ref: '#/components/schemas/Question'
        created_at:
          type: string
          format: date-time
      required: [text, student, question]
...
    Card:
      type: object
      properties:
        id:
          type: integer
          example: 1
        front_text:
          type: string
          example: "Определение предела функции"
        back_text:
          type: string
          example: "Число A называется пределом функции f(x) при x стремящемся к x0..."
        subject:
          type: string
          example: "Математический анализ"
        topic:
          type: string
          example: "Пределы"
      required: [front_text, back_text]
...